+++ src/utils/formatTime.ts
@@ -1,0 +1,3 @@
+export function formatTime(datetime: string): string {
+  return new Date(datetime).getHours().toString();
+}
