+++ src/utils/getWeekday.ts
@@ -1,6 +1,6 @@
 const weekdays = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"] as const;
 
-type Weekday = (typeof weekdays)[number];
+export type Weekday = (typeof weekdays)[number];
 
 export function getWeekday(date: string): Weekday {
   return weekdays[new Date(date).getDay()];
