+++ src/utils/head.ts
@@ -1,3 +1,3 @@
-export function head<T>(array: T[] | string): T | string | undefined {
+export function head<T>(array: T[]): T | undefined {
   return array != null && array.length ? array[0] : undefined;
 }
