+++ src/utils/index.ts
@@ -1,4 +1,7 @@
-export * from "./config";
+export * from "./constants";
 export * from "./head";
-export * from "./transformApiResponse";
 export * from "./getWeekday";
+export * from "./getIconName";
+export * from "./formatTemperature";
+export * from "./formatDate";
+export * from "./formatTime";
